requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.10
requests==2.31.0
boto3==1.34.0
Pillow==10.4.0 